move linux dmesg and audit log collection to conftest
